pyproject.toml
[tool.poetry]
name = "clouddq"
version = "0.3.1"
description = "Data Quality Engine for BigQuery"
authors = ["Thinh Ha <thinhha@google.com>"]
license = "Apache-2.0"

readme = 'README.md'

repository = "https://github.com/GoogleCloudPlatform/cloud-data-quality"
homepage = "https://github.com/GoogleCloudPlatform/cloud-data-quality"

keywords = ['bigquery', 'data quality']

include = ["clouddq/templates/**"]

[tool.poetry.scripts]
clouddq = "clouddq.main:main"

[tool.poetry.dependencies]
python = "^3.8.6,<3.9.0"
click = "^7.1.2"
coloredlogs="^15.0.1"
gitpython="^3.1.24"
Jinja2 = "^2.11.3"
PyYAML = "^5.3.1"
dbt-bigquery = "^0.20.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
black = "^21.5b0"
isort = "^5.6.4"
pyupgrade = "^2.10.0"
flakehell = "^0.9.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true

[tool.flakehell]
pycodestyle = ["+*"]
pyflakes = ["+*"]
format = "grouped"
application_import_names = ['clouddq','tests']
import_order_style = 'google'
docstring_convention = 'google'
max_line_length = 88
max_complexity = 10
show_source = true
extended_default_ignore=[]

[tool.flakehell.plugins]
pyflakes = ["+*"]
"flake8-*" = ["+*"]
pycodestyle = ["+*"]
pylint = ["+F*", "+E*", "-E0611", "-E1101", "-E0401", "-E1102", "-E1123"]
mccabe = ["+*"]
pep8-naming = ["+*"]

[tool.flakehell.exceptions."clouddq/utils.py"]
pyflakes = ['-S404','-S603']

[tool.flakehell.exceptions."tests/*"]
pyflakes = ['-ANN','-S101','-D','-DAR','-S404']

[tool.flakehell.exceptions."**/__init__.py"]
pyflakes = ['-F401']

[tool.isort]
profile = "google"
force_single_line = true
reverse_relative = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["clouddq", "tests"]
skip_glob = ["*/setup.py"]
filter_files = true
known_first_party = "clouddq"
known_third_party = ['click','jinja2','pytest','yaml', 'pyyaml', 'dbt', 'gitpython', 'coloredlogs']
no_lines_before = "LOCALFOLDER"
